Fixed in a kind of bruteforcy way.
At the end of the gameloop a buffer of 10000 space is ensured in the vector containing LogicHubs, meaning that we still incur in the same problem if we spawn more than that in a single frame.
It's not the best but allows us to keep the LogicHubs as normal components and not as a separate system.

Signed-off-by: Andrea Paoletti <andrea.paoletti3@studio.unibo.it>